pyproject.toml
[tool.poetry]
name = "geovision"
version = "0.1.0"
description = ""
authors = ["Sambhav Chandra <vader.sam10@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = {version = "^2.4.0+cu124", source = "pytorch_cu124"}
torchvision = {version = "^0.19.0+cu124", source = "pytorch_cu124"}
transformers = "^4.40.0"
captum = "^0.7.0"
scikit-learn = "^1.4.2"
scikit-image = "^0.23.1"
imageio = "^2.34.0"
imagecodecs = "^2024.1.1"
rasterio = "^1.3.10"
geopandas = "^0.14.3"
fiona = "^1.9.6"
folium = "^0.16.0"
litdata = "^0.2.3"
py7zr = "^0.21.0"
h5py = "^3.11.0"
wandb = "^0.18.0"
awscli = "^1.32.87"
s5cmd = "^0.1.0"
nvitop = "^1.3.2"
jupyterlab = "^4.1.6"
ipykernel = "^6.29.4"
ipympl = "^0.9.4"
python-dotenv = "^1.0.1"
ruff = "0.5.3"
mypy = "^1.9.0"
pydantic = "^2.7.0"
pandera = "^0.20.3"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.29"
pytest = "^8.1.1"
types-requests = "^2.31.0.20240406"
types-tqdm = "^4.66.0.20240417"
aiohttp = "^3.9.5"
aiofiles = "^23.2.1"
types-aiofiles = "^23.2.0.20240403"
tables = "^3.9.2"
memory-profiler = "^0.61.0"
names-generator = "^0.2.0"
jsonargparse = {extras = ["all"], version = "^4.30.0"}
seaborn = "^0.13.2"
geopy = "^2.4.1"
cartopy = "^0.23.0"
openpyxl = "^3.1.5"
timm = "^1.0.9"
lightning = "^2.4.0"
shapely = "^2.0.6"
einops = "^0.8.0"
lxml = "^5.3.0"
html5lib = "^1.1"
rtree = "^1.3.0"

[[tool.poetry.source]]
name = "pytorch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = "*.py"
addopts = "-s -v"

[tool.ruff]
line-length = 150 
indent-width = 4

